+++ awt-blocks.php
@@ -40,6 +40,7 @@ require_once $awt_shared_dir . '/current-url.php';
 require_once $awt_shared_dir . '/faq-schema.php';
 require_once $awt_shared_dir . '/global-controls.php';
 require_once $awt_shared_dir . '/template-chrome.php';
+require_once $awt_shared_dir . '/updates.php';
 unset( $awt_shared_dir );
 
 /**